Refactored add-edit-training component to use directive
