+++ src/app/core/test-utils/MockTrainingService.ts
@@ -1,0 +1,5 @@
+import { Injectable } from '@angular/core';
+import { TrainingService } from '@core/services/training.service';
+
+@Injectable()
+export class MockTrainingService extends TrainingService {}
