+++ src/app/features/workers/add-edit-training/add-edit-training.component.spec.ts
@@ -1,0 +1,60 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { FormBuilder } from '@angular/forms';
+import { ActivatedRoute, RouterModule } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { BreadcrumbService } from '@core/services/breadcrumb.service';
+import { ErrorSummaryService } from '@core/services/error-summary.service';
+import { TrainingService } from '@core/services/training.service';
+import { WorkerService } from '@core/services/worker.service';
+import { MockActivatedRoute } from '@core/test-utils/MockActivatedRoute';
+import { MockBreadcrumbService } from '@core/test-utils/MockBreadcrumbService';
+import { MockTrainingService } from '@core/test-utils/MockTrainingService';
+import { MockWorkerServiceWithWorker } from '@core/test-utils/MockWorkerServiceWithWorker';
+import { SharedModule } from '@shared/shared.module';
+import { getByText, render } from '@testing-library/angular';
+
+import { AddEditTrainingComponent } from './add-edit-training.component';
+
+describe('AddEditTrainingComponent', () => {
+  async function setup() {
+    const { fixture, getByText } = await render(AddEditTrainingComponent, {
+      imports: [SharedModule, RouterModule, RouterTestingModule, HttpClientTestingModule],
+      providers: [
+        { provide: BreadcrumbService, useClass: MockBreadcrumbService },
+        {
+          provide: ActivatedRoute,
+          useValue: new MockActivatedRoute({
+            snapshot: {
+              params: { trainingRecordId: '1' },
+            },
+            parent: {
+              snapshot: {
+                data: {
+                  establishment: {
+                    uid: '1',
+                  },
+                },
+              },
+            },
+          }),
+        },
+        FormBuilder,
+        ErrorSummaryService,
+        { provide: TrainingService, useClass: MockTrainingService },
+        { provide: WorkerService, useClass: MockWorkerServiceWithWorker },
+      ],
+    });
+
+    const component = fixture.componentInstance;
+    return {
+      component,
+      fixture,
+      getByText,
+    };
+  }
+
+  it('should create', async () => {
+    const { component } = await setup();
+    expect(component).toBeTruthy();
+  });
+});
